* refactoring of form models * sorting data from backend * fix some bugs
