+++ src/App.tsx
@@ -20,43 +20,35 @@ const loginNeededText = 'Musisz być zalogowany!';
 
 const AppInternal = () => {
     return (
-        <>
-            <Switch>
-                <PrivateRoute exact path={'/'} component={HomePageComponent} redirectPath={'/login'} reverse />
+        <Switch>
+            <PrivateRoute exact path={'/'} component={HomePageComponent} redirectPath={'/login'} reverse />
 
-                <PrivateRoute path={'/login'} exact component={LoginComponent} redirectPath={'/dashboard'} reverse />
+            <PrivateRoute path={'/login'} exact component={LoginComponent} redirectPath={'/dashboard'} reverse />
 
-                <PrivateRoute
-                    path={'/register'}
-                    exact
-                    component={RegisterComponent}
-                    redirectPath={'/dashboard'}
-                    reverse
-                />
+            <PrivateRoute path={'/register'} exact component={RegisterComponent} redirectPath={'/dashboard'} reverse />
 
-                <Route exact path={'/dashboard'} component={DashboardComponent} />
+            <Route exact path={'/dashboard'} component={DashboardComponent} />
 
-                <Route exact path={'/articles'} component={ArticleListContainer} />
-                <Route exact path={'/articles/:id'} component={SingleArticleComponent} />
+            <Route exact path={'/articles'} component={ArticleListContainer} />
+            <Route exact path={'/articles/:id'} component={SingleArticleComponent} />
 
-                <Route exact path={'/profile/:id'} component={UserProfileComponent} />
-                <PrivateRoute
-                    path={'/settings'}
-                    exact
-                    component={SettingsComponent}
-                    redirectPath={'/login'}
-                    redirectText={loginNeededText}
-                />
+            <Route exact path={'/profile/:id'} component={UserProfileComponent} />
+            <PrivateRoute
+                path={'/settings'}
+                exact
+                component={SettingsComponent}
+                redirectPath={'/login'}
+                redirectText={loginNeededText}
+            />
 
-                <PrivateRoute
-                    path={'/tasks'}
-                    exact
-                    component={TaskListComponent}
-                    redirectPath={'/login'}
-                    redirectText={loginNeededText}
-                />
-            </Switch>
-        </>
+            <PrivateRoute
+                path={'/tasks'}
+                exact
+                component={TaskListComponent}
+                redirectPath={'/login'}
+                redirectText={loginNeededText}
+            />
+        </Switch>
     );
 };
