+++ src/components/articles/list/articleCategoryComponent.tsx
@@ -15,13 +15,15 @@ const ArticleCategoryComponent = ({ category }: ArticleCategoryProps) => {
             <CategoryHeader>{category.name}</CategoryHeader>
             <ArticleCategoryDesc>{category.description}</ArticleCategoryDesc>
             <ArticlesCategoryCardsContainer>
-                {category.articles.map((a) => {
-                    return (
-                        <div className="w-full md:w-1/2 lg:w-3/12 py-4 px-2">
-                            <ArticleCard article={a} key={a.id} />
-                        </div>
-                    );
-                })}
+                {category.articles
+                    .sort((a1, a2) => a1.ordinal - a2.ordinal)
+                    .map((a) => {
+                        return (
+                            <div key={a.id} className="w-full md:w-1/2 lg:w-3/12 py-4 px-2">
+                                <ArticleCard article={a} />
+                            </div>
+                        );
+                    })}
             </ArticlesCategoryCardsContainer>
         </>
     );
