+++ src/components/articles/list/articleListContainer.tsx
@@ -14,9 +14,11 @@ const ArticleListContainer = () => {
         <ComponentStateHandler state={state}>
             <CardsContainer>
                 {articleCategory &&
-                    articleCategory.map((ac) => {
-                        return <ArticleCategoryComponent key={ac.id} category={ac} />;
-                    })}
+                    articleCategory
+                        .sort((c1, c2) => c1.ordinal - c2.ordinal)
+                        .map((ac) => {
+                            return <ArticleCategoryComponent key={ac.id} category={ac} />;
+                        })}
             </CardsContainer>
         </ComponentStateHandler>
     );
