+++ src/components/login/loginComponent.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { LoginData } from '../../models/loginData';
 import { authService } from '../../services/authService';
 import LoginForm from './loginForm';
 import { useToasts } from 'react-toast-notifications';
@@ -9,6 +8,7 @@ import { User } from '../../models/user';
 import { loginFailure, loginSuccess } from '../../context/userReducer';
 import { formTexts } from '../shared/form/formTexts';
 import { AxiosError } from 'axios';
+import { LoginData } from '../../models/formData';
 
 const LoginComponent = () => {
     const [loading, setLoading] = useState(false);
