+++ src/components/login/loginForm.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useForm } from 'react-hook-form';
-import { LoginData } from '../../models/loginData';
 import FormInputField from '../shared/form/formInputField';
 import { formTexts } from '../shared/form/formTexts';
 import SubmitButton from '../shared/form/submitButton';
 import FormHeaderComponent from '../shared/form/formHeader';
 import { handleBackendErrors } from '../shared/form/backendErrorHandler';
+import { LoginData } from '../../models/formData';
 
 interface LoginFormProps {
     loading: boolean;
