+++ src/components/register/registerComponent.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { authService } from '../../services/authService';
 import RegisterForm from './registerForm';
-import { RegisterData } from '../../models/registerData';
 import { useToasts } from 'react-toast-notifications';
 import { useAppDispatch } from '../../context/hooks';
 import { useHistory } from 'react-router-dom';
@@ -9,6 +8,7 @@ import { registerFailure, registerSuccess } from '../../context/userReducer';
 import { User } from '../../models/user';
 import { formTexts } from '../shared/form/formTexts';
 import { AxiosError } from 'axios';
+import { RegisterData } from '../../models/formData';
 
 const RegisterComponent = () => {
     const [loading, setLoading] = useState(false);
