+++ src/components/register/registerForm.tsx
@@ -1,11 +1,11 @@
 import React, { useRef } from 'react';
 import { useForm } from 'react-hook-form';
-import { RegisterData } from '../../models/registerData';
 import { formTexts } from '../shared/form/formTexts';
 import FormInputField from '../shared/form/formInputField';
 import FormHeaderComponent from '../shared/form/formHeader';
 import SubmitButton from '../shared/form/submitButton';
 import { handleBackendErrors } from '../shared/form/backendErrorHandler';
+import { RegisterData } from '../../models/formData';
 
 interface RegisterFormProps {
     loading: boolean;
