+++ src/components/settings/deleteButton.tsx
@@ -13,8 +13,8 @@ import { profileService } from '../../services/profileService';
 import { useHistory } from 'react-router-dom';
 import { useAppDispatch } from '../../context/hooks';
 import { useToasts } from 'react-toast-notifications';
-import { DeleteAccountData } from '../../models/deleteAccountData';
 import { handleBackendErrors } from '../shared/form/backendErrorHandler';
+import { DeleteAccountData } from '../../models/formData';
 
 interface DeleteButtonProps {}
