+++ src/components/settings/settingsComponent.tsx
@@ -6,9 +6,9 @@ import { User } from '../../models/user';
 import { formTexts } from '../shared/form/formTexts';
 import { AxiosError } from 'axios';
 import SettingsForm from './settingsForm';
-import { SettingsData } from '../../models/settingsData';
 import { profileService } from '../../services/profileService';
 import { useHistory } from 'react-router-dom';
+import { SettingsData } from '../../models/formData';
 
 const SettingsComponent = () => {
     const [loading, setLoading] = useState(false);
