+++ src/components/settings/settingsForm.tsx
@@ -4,11 +4,11 @@ import { formTexts } from '../shared/form/formTexts';
 import FormInputField from '../shared/form/formInputField';
 import FormHeaderComponent from '../shared/form/formHeader';
 import SubmitButton from '../shared/form/submitButton';
-import { SettingsData } from '../../models/settingsData';
 import { useAppSelector } from '../../context/hooks';
 import { currentUserData } from '../../context/userReducer';
 import DeleteAccountButton from './deleteButton';
 import { handleBackendErrors } from '../shared/form/backendErrorHandler';
+import { SettingsData } from '../../models/formData';
 
 interface RegisterFormProps {
     loading: boolean;
