+++ src/components/tasks/list/taskCategoryComponent.tsx
@@ -18,7 +18,7 @@ const TaskCategoryComponent = ({ category }: TaskCategoryProps) => {
                     .map((task) => {
                         return (
                             <div key={task.id} className="flex justify-center w-full md:w-1/2 lg:w-1/3 py-4 px-4">
-                                <TaskCard key={task.id} taskData={task} />
+                                <TaskCard taskData={task} />
                             </div>
                         );
                     })}
