+++ src/components/tasks/single/singleTaskContent.tsx
@@ -4,7 +4,6 @@ import { markdown } from '../../../remarkable/remarkable';
 import { Task, TaskStatus } from '../../../models/task';
 import SingleTaskContainer from '../../styled-components/tasks/SingleTaskContainer';
 import { AttachmentModel } from '../../../models/attachment';
-import { SolveData } from '../../../models/solveData';
 import { taskService } from '../../../services/taskService';
 import SolveForm from './solveForm';
 import { wrapTags } from '../../shared/tags/wrapTags';
@@ -15,6 +14,7 @@ import { useAppDispatch } from '../../../context/hooks';
 import { solveTask } from '../../../context/userReducer';
 import TaskSolvedBanner from './taskSolvedBanner';
 import Attachment from '../../remarkable-components/attachment';
+import { SolveData } from '../../../models/formData';
 
 interface SingleTaskProps {
     task: Task;
