+++ src/components/tasks/single/solveForm.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useForm } from 'react-hook-form';
-import { SolveData } from '../../../models/solveData';
 import { formTexts } from '../../shared/form/formTexts';
 import FormInputField from '../../shared/form/formInputField';
 import SubmitButton from '../../shared/form/submitButton';
 import { handleBackendErrors } from '../../shared/form/backendErrorHandler';
+import { SolveData } from '../../../models/formData';
 
 interface SolveFormProps {
     loading: boolean;
