+++ src/models/article.ts
@@ -3,7 +3,7 @@ import { TaskCardData } from './task';
 export interface Article {
     id: number;
     name: string;
-    oridinal: number;
+    ordinal: number;
     category: number;
     content: string;
     excerpt: string;
@@ -13,6 +13,7 @@ export interface Article {
 
 export interface ArticleCategory {
     id: number;
+    ordinal: number;
     name: string;
     description: string;
     articles: Article[];
