--- src/models/deleteAccountData.ts
@@ -1,3 +1,0 @@
-export interface DeleteAccountData {
-    password: string;
-}
