+++ src/models/formData.ts
@@ -1,0 +1,34 @@
+export interface DeleteAccountData {
+    password: string;
+}
+
+export interface LoginData {
+    username: string;
+    password: string;
+}
+
+export interface RegisterData {
+    username: string;
+    email: string;
+    password: string;
+    passwordConfirmation: string;
+}
+
+export interface ResponseFlagData {
+    id: number;
+    flag: string;
+    success: boolean;
+}
+
+export interface SettingsData {
+    email: string;
+    newPassword: string;
+    passwordConfirmation: string;
+
+    oldPassword: string;
+}
+
+export interface SolveData {
+    id: number;
+    flag: string;
+}
