--- src/models/loginData.ts
@@ -1,4 +1,0 @@
-export interface LoginData {
-    username: string;
-    password: string;
-}
