--- src/models/registerData.ts
@@ -1,6 +1,0 @@
-export interface RegisterData {
-    username: string;
-    email: string;
-    password: string;
-    passwordConfirmation: string;
-}
