--- src/models/responseFlagData.ts
@@ -1,5 +1,0 @@
-export interface ResponseFlagData {
-    id: number;
-    flag: string;
-    success: boolean;
-}
