--- src/models/settingsData.ts
@@ -1,7 +1,0 @@
-export interface SettingsData {
-    email: string;
-    newPassword: string;
-    passwordConfirmation: string;
-
-    oldPassword: string;
-}
