--- src/models/solveData.ts
@@ -1,4 +1,0 @@
-export interface SolveData {
-    id: number;
-    flag: string;
-}
