+++ src/services/authService.ts
@@ -1,12 +1,11 @@
 import { User } from '../models/user';
-import { LoginData } from '../models/loginData';
-import { RegisterData } from '../models/registerData';
 import axiosInstance from '../utils/axiosInterceptors';
 import { Token } from '../models/token';
 import { apiUrls } from '../utils/apiUrls';
 import { AxiosError, AxiosResponse } from 'axios';
 import { store } from '../context/store';
 import { updateTokens } from '../context/userReducer';
+import { LoginData, RegisterData } from '../models/formData';
 
 const responseBody = (response: AxiosResponse) => response.data;
