+++ src/services/profileService.ts
@@ -2,12 +2,11 @@ import axiosInstance from '../utils/axiosInterceptors';
 import { AxiosRequestConfig, AxiosResponse } from 'axios';
 import { apiUrls } from '../utils/apiUrls';
 import { User, UserData } from '../models/user';
-import { SettingsData } from '../models/settingsData';
-import { DeleteAccountData } from '../models/deleteAccountData';
+import { DeleteAccountData, SettingsData } from '../models/formData';
 
 const responseBody = (response: AxiosResponse) => response.data;
 
-const get = async (id: string, options? :AxiosRequestConfig): Promise<UserData> => {
+const get = async (id: string, options?: AxiosRequestConfig): Promise<UserData> => {
     const url = `${apiUrls.USERS}${id}/`;
     return axiosInstance.get(url, options).then(responseBody);
 };
