+++ src/services/taskService.ts
@@ -2,16 +2,15 @@ import { AxiosRequestConfig, AxiosResponse } from 'axios';
 import axiosInstance from '../utils/axiosInterceptors';
 import { Task, TaskCardData } from '../models/task';
 import { apiUrls } from '../utils/apiUrls';
-import { SolveData } from '../models/solveData';
-import { ResponseFlagData } from '../models/responseFlagData';
+import { ResponseFlagData, SolveData } from '../models/formData';
 
 const responseBody = (response: AxiosResponse) => response.data;
 
-const getAll = async (options? :AxiosRequestConfig): Promise<TaskCardData[]> => {
+const getAll = async (options?: AxiosRequestConfig): Promise<TaskCardData[]> => {
     return axiosInstance.get(apiUrls.TASKS, options).then(responseBody);
 };
 
-const getOne = async (id: number, options? :AxiosRequestConfig): Promise<Task> => {
+const getOne = async (id: number, options?: AxiosRequestConfig): Promise<Task> => {
     const url = `${apiUrls.TASKS}${id}/`;
     return axiosInstance.get(url, options).then(responseBody);
 };
